design/#13 뷰 다 그리고 데이터 수정하고 저장까지 됨. 이제 삭제 기능하고 자잘한 버그들 수정해야됨
