+++ Pikacho/Pikacho/View/CheckView.swift
@@ -76,6 +76,7 @@ struct CheckView: View {
                             }
                     }
                 }
+
     }
 
     private func loadBookings() {
